requirements.txt
gps3
paho-mqtt
haversine